+++ app/Exceptions/Social/IncompleteMetaGraphPaginationException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services\Social\Meta;
+namespace App\Exceptions\Social;
 
 use RuntimeException;
 use Throwable;
@@ -11,7 +11,7 @@ use Throwable;
  * Thrown when a Meta Graph edge was only partially fetched. Callers must not
  * treat a truncated list as complete (e.g. auto-connect when count === 1).
  */
-class IncompletePaginationException extends RuntimeException
+class IncompleteMetaGraphPaginationException extends RuntimeException
 {
     public function __construct(?Throwable $previous = null)
     {
